backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-multipart==0.0.9
pandas
pyarrow
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic
pydantic-settings
httpx==0.27.0
scipy
numpy
requests==2.31.0
groq
